Create BaseGame class with tick and deltaTime capability.
Why: Managing time is one of the most important parts in game.
Especially this game engine is for real-time based, not turn based.
Capabilities added with this commit will be core of game, so I developed
this capabilities first.

Description: When game starts, Tick() function will be called in every
loop. In every tick, deltaTime will be updated. This time is calculated
based on clock counts and CLOCK_PER_SEC in <ctime> header.

Tags: Add